Add in support for f2fs
Was having issues with getting the rom to boot so looking at the source tree
of ShevT for lt033g device was able to observe the correct mounting flags
for userdata and cache partitions. Once these were moved across, these
partions began to function. In relation to system i simply added this
and removed extra flags causing no boot.

Change-Id: I4a4a00624e9a2c71e62e18d496de44fe2848dc30
Signed-off-by: Joshndroid <fmxjoshiewah@gmail.com>